Add files via upload
The rise of electric vehicles has resulted in a clear decline in gasoline sales. This reduction in consumption is a positive step towards reducing the environmental impact of transportation. The growth of the EV market is contributing to the development of new technologies and infrastructure for cleaner forms of transportation.